@@ -1,10 +1,8 @@
 package com.snowball.backend.entity;
 
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
-
-import javax.persistence.*;
-import java.util.Date;
 
 @Entity
 @Getter
